@@ -1,6 +1,6 @@
 module doh
 
-go 1.20
+go 1.21
 
 require (
 	github.com/miekg/dns v1.1.55
@@ -9,7 +9,7 @@ require (
 
 require (
 	golang.org/x/mod v0.12.0 // indirect
-	golang.org/x/net v0.12.0 // indirect
-	golang.org/x/sys v0.10.0 // indirect
-	golang.org/x/tools v0.11.0 // indirect
+	golang.org/x/net v0.14.0 // indirect
+	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/tools v0.12.0 // indirect
 )
